Move posts processing logic to PostsProcessor
