+++ tg-bot/src/main/java/polis/posting/IPostsProcessor.java
@@ -1,0 +1,9 @@
+package polis.posting;
+
+import org.telegram.telegrambots.meta.api.objects.Message;
+
+import java.util.List;
+
+public interface IPostsProcessor {
+    void processPostsInChannel(long channelId, List<Message> posts);
+}
